+++ internal/batches/docker/image.go
@@ -66,19 +66,47 @@ func (image *image) Ensure(ctx context.Context) error {
 	image.ensureOnce.Do(func() {
 		image.ensureErr = func() (err error) {
 			inspectDigest := func() (string, error) {
-				// TODO!(sqs): is image id the right thing to use here? it is NOT
-				// the digest. but the digest is not calculated for all images
-				// (unless they are pulled/pushed from/to a registry), see
-				// https://github.com/moby/moby/issues/32016.
-				out, err := exec.CommandContext(ctx, "docker", "image", "inspect", "--format", "{{ .Id }}", image.name).CombinedOutput()
+				// Since we are only asking Docker for local information, we
+				// expect this operation to be quick, and therefore set a
+				// relatively low timeout for Docker to respond. This is
+				// particularly useful because this function is usually the
+				// first non-trivial interaction we have with Docker in a
+				// src-cli invocation, and this allows us to catch failure modes
+				// that result in the Docker socket still listening and
+				// accepting connections, but where dockerd is no longer able to
+				// respond to non-trivial requests.
+				//
+				// Anecdotally, this seems to happen most frequently with Docker
+				// Desktop VMs running out of memory, whereupon the Linux
+				// kernel's OOM killer sometimes chooses to kill components of
+				// Docker instead of processes within containers.
+				dctx, cancel, err := withFastCommandContext(ctx)
+				if err != nil {
+					return "", err
+				}
+				defer cancel()
+
+				args := []string{"image", "inspect", "--format", "{{ .Id }}", image.name}
+				out, err := exec.CommandContext(dctx, "docker", args...).CombinedOutput()
 				id := string(bytes.TrimSpace(out))
-				return id, err
+
+				if errors.IsDeadlineExceeded(err) || errors.IsDeadlineExceeded(dctx.Err()) {
+					return "", newFastCommandTimeoutError(dctx, args...)
+				} else if err != nil {
+					return "", err
+				}
+
+				return id, nil
 			}
 
 			// docker image inspect will return a non-zero exit code if the image and
 			// tag don't exist locally, regardless of the format.
 			var digest string
-			if digest, err = inspectDigest(); err != nil {
+			if digest, err = inspectDigest(); errors.HasType(err, &fastCommandTimeoutError{}) {
+				// Ensure we immediately propagate a timeout up, rather than
+				// trying to tell an unresponsive Docker to pull.
+				return err
+			} else if err != nil {
 				// Let's try pulling the image.
 				if err := exec.CommandContext(ctx, "docker", "image", "pull", image.name).Run(); err != nil {
 					return errors.Wrap(err, "pulling image")