+++ internal/batches/ui/json_lines.go
@@ -94,8 +94,8 @@ func (ui *JSONLines) CheckingCacheSuccess(cachedSpecsFound int, tasksToExecute i
 	})
 }
 
-func (ui *JSONLines) ExecutingTasks(verbose bool, parallelism int) executor.TaskExecutionUI {
-	return &taskExecutionJSONLines{verbose: verbose, parallelism: parallelism}
+func (ui *JSONLines) ExecutingTasks(_ bool, _ int) executor.TaskExecutionUI {
+	return &taskExecutionJSONLines{}
 }
 
 func (ui *JSONLines) ExecutingTasksSkippingErrors(err error) {
@@ -175,10 +175,7 @@ func (ui *JSONLines) ExecutionError(err error) {
 var _ executor.JSONCacheWriter = &JSONLines{}
 
 func (ui *JSONLines) WriteExecutionResult(key string, value execution.Result) {
-	logOperationSuccess(batcheslib.LogEventOperationCacheResult, &batcheslib.CacheResultMetadata{
-		Key:   key,
-		Value: value,
-	})
+	// Noop, we don't need execution results.
 }
 
 func (ui *JSONLines) WriteAfterStepResult(key string, value execution.AfterStepResult) {
@@ -189,9 +186,6 @@ func (ui *JSONLines) WriteAfterStepResult(key string, value execution.AfterStepR
 }
 
 type taskExecutionJSONLines struct {
-	verbose     bool
-	parallelism int
-
 	linesTasks map[*executor.Task]batcheslib.JSONLinesTask
 }
 
@@ -286,21 +280,17 @@ type stepsExecutionJSONLines struct {
 const stepFlushDuration = 500 * time.Millisecond
 
 func (ui *stepsExecutionJSONLines) ArchiveDownloadStarted() {
-	logOperationStart(batcheslib.LogEventOperationTaskDownloadingArchive, &batcheslib.TaskDownloadingArchiveMetadata{TaskID: ui.linesTask.ID})
+	// We don't fetch archives in executor mode.
 }
 func (ui *stepsExecutionJSONLines) ArchiveDownloadFinished(err error) {
-	if err != nil {
-		logOperationFailure(batcheslib.LogEventOperationTaskDownloadingArchive, &batcheslib.TaskDownloadingArchiveMetadata{TaskID: ui.linesTask.ID, Error: err.Error()})
-	} else {
-		logOperationSuccess(batcheslib.LogEventOperationTaskDownloadingArchive, &batcheslib.TaskDownloadingArchiveMetadata{TaskID: ui.linesTask.ID})
-	}
+	// We don't fetch archives in executor mode.
 }
 
 func (ui *stepsExecutionJSONLines) WorkspaceInitializationStarted() {
-	logOperationStart(batcheslib.LogEventOperationTaskInitializingWorkspace, &batcheslib.TaskInitializingWorkspaceMetadata{TaskID: ui.linesTask.ID})
+	// No workspace initialization required for executor mode.
 }
 func (ui *stepsExecutionJSONLines) WorkspaceInitializationFinished() {
-	logOperationSuccess(batcheslib.LogEventOperationTaskInitializingWorkspace, &batcheslib.TaskInitializingWorkspaceMetadata{TaskID: ui.linesTask.ID})
+	// No workspace initialization required for executor mode.
 }
 
 func (ui *stepsExecutionJSONLines) SkippingStepsUpto(startStep int) {
@@ -361,13 +351,6 @@ func (ui *stepsExecutionJSONLines) StepFailed(step int, err error, exitCode int)
 			ExitCode: exitCode,
 		},
 	)
-}
-
-func (ui *stepsExecutionJSONLines) CalculatingDiffStarted() {
-	logOperationStart(batcheslib.LogEventOperationTaskCalculatingDiff, &batcheslib.TaskCalculatingDiffMetadata{TaskID: ui.linesTask.ID})
-}
-func (ui *stepsExecutionJSONLines) CalculatingDiffFinished() {
-	logOperationSuccess(batcheslib.LogEventOperationTaskCalculatingDiff, &batcheslib.TaskCalculatingDiffMetadata{TaskID: ui.linesTask.ID})
 }
 
 func logOperationStart(op batcheslib.LogEventOperation, metadata interface{}) {