+++ internal/batches/workspace/git.go
@@ -15,6 +15,8 @@ func runGitCmd(ctx context.Context, dir string, args ...string) ([]byte, error) 
 		"GIT_CONFIG_NOSYSTEM=1",
 		// And also not any other, because they can mess up output, change defaults, .. which can do unexpected things.
 		"GIT_CONFIG=/dev/null",
+		// Don't ask interactively for credentials.
+		"GIT_TERMINAL_PROMPT=0",
 		// Set user.name and user.email in the local repository. The user name and
 		// e-mail will eventually be ignored anyway, since we're just using the Git
 		// repository to generate diffs, but we don't want git to generate alarming
@@ -25,9 +27,12 @@ func runGitCmd(ctx context.Context, dir string, args ...string) ([]byte, error) 
 		"GIT_COMMITTER_EMAIL=batch-changes@sourcegraph.com",
 	}
 	cmd.Dir = dir
-	out, err := cmd.CombinedOutput()
+	out, err := cmd.Output()
 	if err != nil {
-		return nil, errors.Wrapf(err, "'git %s' failed: %s", strings.Join(args, " "), out)
+		if exitErr, ok := err.(*exec.ExitError); ok {
+			return out, errors.Wrapf(err, "'git %s' failed: %s", strings.Join(args, " "), string(exitErr.Stderr))
+		}
+		return out, errors.Wrapf(err, "'git %s' failed: %s", strings.Join(args, " "), string(out))
 	}
 	return out, nil
 }